refactor(github-pulls): modularize pull commands
